refactor: renamed component
OverviewActions is renamed to OverviewDropDownActions
